+++ src/components/trial_screens/SelfPacedReadingScreen.vue
@@ -71,7 +71,7 @@ import SelfPacedReadingInput from '../inputs/SelfPacedReadingInput';
  * |Measurement|Type|Description|
  * |--|---|---|
  * |responseTime|int|Time between display of this screen until the screen is left|
- * |response_timeout|boolean|Whether the response time limit was reached|
+ * |responseTimeout|boolean|Whether the response time limit was reached|
  * |qud *(optional)*|string||
  * |pauseTime *(optional)*|int||
  * |fixationTime *(optional)*|int||
@@ -147,7 +147,7 @@ export default {
   data() {
     return {
       responseTimes: [],
-      response_timeout: false
+      responseTimeout: false
     };
   }
 };